Added support for Zinc 1.2.x
